+++ packages/read-models/src/index.ts
@@ -17,3 +17,4 @@ export * from "./player-priors.js";
 export * from "./birth-notice-articles.js";
 export * from "./news-articles.js";
 export * from "./life-dossier.js";
+export * from "./player-articles.js";